Añadido movimiento de las basuras B en el eje X. Falta colocar el movimiento en el eje, renombrar los movimientos y corregir el movimiento táctil.

+++ app/src/main/java/mx/a01736353/savethebunny/Explosion.java
@@ -7,7 +7,7 @@ import android.graphics.BitmapFactory;
 public class Explosion {
     Bitmap explosion[] = new Bitmap[4];
     int explosionFrame = 0;
-    int explosionX, explosionY;
+    float explosionX, explosionY;
 
     //Constructor
     public Explosion(Context context){
